removed superfluos cse/indexing/posting and moved classes to cse/indexing
