--- cse/indexing/posting/__init__.py
@@ -1,8 +1,0 @@
-"""
-CSE - a web crawling and web searching application for news paper
-comments written in Python
-
-package for the default inverted posting lists index consisting of
- - PostingList object definition with decoder and encoder
- - custom index writer logic
-"""
